+++ src/Figuri/PyatiUgol.java
@@ -6,8 +6,8 @@ public class PyatiUgol extends Figura {
     }
 
     @Override
-    public int poiskS() {
-        return 0;
+    public Double poiskS() {
+        return 0.0;
     }
 
     @Override
